pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contrib-checker"
version = "1.0.0"
description = "Check if code contributors are properly listed in metadata files such as CITATION.cff and codemeta.json based on the git history"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Thomas Vuillaume", email = "thomas.vuillaume@lapp.in2p3.fr"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Documentation",
]
keywords = ["git", "contributors", "citation", "metadata", "github-actions", "gitlab-ci"]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=5.4.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
]

[project.urls]
Homepage = "https://github.com/vuillaut/contrib-checker"
Repository = "https://github.com/vuillaut/contrib-checker"
Issues = "https://github.com/vuillaut/contrib-checker/issues"

[project.scripts]
contrib-checker = "contrib_checker.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["contrib_checker*"]

[tool.ruff]
exclude = [
    ".eggs",
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "site-packages",
]
line-length = 99
lint.select = ["E4", "E7", "E9", "F"]
lint.ignore = []
lint.fixable = ["ALL"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
